+++ src/OpenTelemetry.Api/Context/Propagation/B3Propagator.cs
@@ -26,6 +26,7 @@ namespace OpenTelemetry.Context.Propagation
     /// <summary>
     /// A text map propagator for B3. See https://github.com/openzipkin/b3-propagation.
     /// </summary>
+    [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
     public sealed class B3Propagator : TextMapPropagator
     {
         internal const string XB3TraceId = "X-B3-TraceId";
@@ -58,6 +59,7 @@ namespace OpenTelemetry.Context.Propagation
         /// <summary>
         /// Initializes a new instance of the <see cref="B3Propagator"/> class.
         /// </summary>
+        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
         public B3Propagator()
             : this(false)
         {
@@ -67,6 +69,7 @@ namespace OpenTelemetry.Context.Propagation
         /// Initializes a new instance of the <see cref="B3Propagator"/> class.
         /// </summary>
         /// <param name="singleHeader">Determines whether to use single or multiple headers when extracting or injecting span context.</param>
+        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
         public B3Propagator(bool singleHeader)
         {
             this.singleHeader = singleHeader;
@@ -76,7 +79,10 @@ namespace OpenTelemetry.Context.Propagation
         public override ISet<string> Fields => AllFields;
 
         /// <inheritdoc/>
+        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+#pragma warning disable CS0809 // Obsolete member overrides non-obsolete member
         public override PropagationContext Extract<T>(PropagationContext context, T carrier, Func<T, string, IEnumerable<string>> getter)
+#pragma warning restore CS0809 // Obsolete member overrides non-obsolete member
         {
             if (context.ActivityContext.IsValid())
             {
@@ -107,7 +113,10 @@ namespace OpenTelemetry.Context.Propagation
         }
 
         /// <inheritdoc/>
+        [Obsolete("Use OpenTelemetry.Extensions.Propagators package instead.")]
+#pragma warning disable CS0809 // Obsolete member overrides non-obsolete member
         public override void Inject<T>(PropagationContext context, T carrier, Action<T, string, string> setter)
+#pragma warning restore CS0809 // Obsolete member overrides non-obsolete member
         {
             if (context.ActivityContext.TraceId == default || context.ActivityContext.SpanId == default)
             {